muestra nulos y las tablas error con mas de una pestaña

+++ PArticulo/PArticulo/Categoria.cs
@@ -4,9 +4,9 @@ namespace PArticulo
 {
 	public class Categoria
 	{
-		public int id;
+		public UInt64 id;
 		public string nombre;
-		public Categoria (int id,string nombre)
+		public Categoria (UInt64 id,string nombre)
 		{
 			this.id = id;
 			this.nombre = nombre;
@@ -14,4 +14,3 @@ namespace PArticulo
 
 	}
 }
-
